+++ src/DotNetProjectFile.Analyzers/DotNetProjectFile.Analyzers.csproj
@@ -28,7 +28,7 @@
     <PackageReleaseNotes>
       <![CDATA[
 ToBeReleased
-- Proj0030: Use VB.NET specific properties only in when applicable. (NEW RULE)
+- Proj0030: Use VB.NET specific properties only when applicable. (NEW RULE)
 - Proj0216: Define the product name explictly. (NEW RULE)
 v1.4.7
 - Proj2005: Escape XML nodes of resource values. (NEW RULE)
