+++ src/DotNetProjectFile.Analyzers/MsBuild/VisualBasic/NoVBRuntimeReference.cs
@@ -1,4 +1,4 @@
-namespace DotNetProjectFile.MsBuild;
+namespace DotNetProjectFile.MsBuild.VisualBasic;
 
 /// <summary>
 /// A boolean value that indicates whether the Visual Basic runtime (Microsoft.VisualBasic.dll)
