+++ src/DotNetProjectFile.Analyzers/MsBuild/VisualBasic/OptionExplicit.cs
@@ -1,4 +1,4 @@
-namespace DotNetProjectFile.MsBuild;
+namespace DotNetProjectFile.MsBuild.VisualBasic;
 
 /// <summary>
 /// A boolean value that when set to true, requires explicit declaration of
