+++ src/DotNetProjectFile.Analyzers/MsBuild/VisualBasic/OptionInfer.cs
@@ -1,4 +1,4 @@
-namespace DotNetProjectFile.MsBuild;
+namespace DotNetProjectFile.MsBuild.VisualBasic;
 
 /// <summary>
 /// A boolean value that when set to true, enables type inference of variables.
