+++ src/DotNetProjectFile.Analyzers/Rule.cs
@@ -256,7 +256,7 @@ public static class Rule
 
     public static DiagnosticDescriptor UseInVBContextOnly => New(
         id: 0030,
-        title: "Use VB.NET specific properties only in when applicable",
+        title: "Use VB.NET specific properties only when applicable",
         message: "The property <{0}> is only applicable when using VB.NET and can therefor be removed.",
         description:
             "Properties only applicable to VB.NET are noise when none of " +
